nats-test/requirements.txt
nats-py==2.1.7